requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.0
kiwisolver==1.5.0
matplotlib==3.10.8
mss==10.1.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pure-python-adb==0.3.0.dev0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
stable_baselines3
tensorboard
rich